File structure change: moved util classes to their own folder to keep tidy
