+++ BuldingMapper/Controls/FloorEditor.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using BuildingMapper.FormResults;
+using BuildingMapper.Util;
 
 namespace BuildingMapper
 {
