+++ BuldingMapper/Util/Floor.cs
@@ -4,29 +4,29 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BuildingMapper
+namespace BuildingMapper.Util
 {
     public class Floor
     {
         public string Name { get; set; }
         public List<Room> Rooms { get; set; }
 
-        public Floor(string name) 
-        { 
+        public Floor(string name)
+        {
             Name = name;
             Rooms = new List<Room>();
         }
 
         public Floor(string name, List<Room> rooms) : this(name)
         {
-            this.Rooms = rooms;
+            Rooms = rooms;
         }
 
         public static List<Floor> ConvertToFloorObjects(List<FloorEditor> floorEditors)
         {
             List<Floor> result = new List<Floor>();
 
-            foreach(FloorEditor f in floorEditors)
+            foreach (FloorEditor f in floorEditors)
             {
                 result.Add(new Floor(f.FloorName, f.Rooms));
             }
